requirements.txt
streamlit
python-dotenv
playwright
langchain-ollama
langchain-core
openai